requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
beautifulsoup4
chromadb